Move the DEFAULT_CONFIG
